requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
psycopg[binary]>=3.2,<4.0
PyJWT>=2.9,<3.0
pydantic>=2.8,<3.0